refactor(timelapse): periodic_merge.go 按域拆分 — 1432 行 → 5 个按域文件 (#611)
纯移动无逻辑变化：periodic_merge.go(572, 类型/选项/构造 + Run 管线) + periodic_merge_strategies.go(393, go/ffmpeg concat + goMergeSegments + 分 codec 合并) + periodic_merge_records.go(234, finalizeMerge/timelapse_merges 行写入/中间 mp4 清理) + periodic_merge_status.go(202, 失败标记/进度回写/段过滤) + periodic_merge_probe.go(278, 兼容性检查/元数据与帧数探测/录像帧提取)
